+++ src/FieldWarning/Assets/Units/Component/Weapon/TurretSystem.cs
@@ -49,9 +49,9 @@ namespace PFW.Units.Component.Weapon
         }
 
         /// <summary>
-        /// The max fire range to the current target.
+        /// The max fire range to the current target, in unity units.
         /// </summary>
-        private int _fireRange;  
+        private float _fireRange;  
         public List<Turret> Children; // sync
 
         /// <summary>
@@ -229,12 +229,12 @@ namespace PFW.Units.Component.Weapon
         /// TODO: Code duplication can be reduced if we only implement this in 
         /// the turret class and have a fake toplevel turret we call this method on,
         /// but a fake turret like that also adds complexity, hard to decide.
-        private int MaxRange(TargetTuple target)
+        private float MaxRange(TargetTuple target)
         {
-            int maxRange = 0;
+            float maxRange = 0;
             foreach (Turret turret in Children)
             {
-                int turretMax = turret.MaxRange(target);
+                float turretMax = turret.MaxRange(target);
                 maxRange = maxRange > turretMax ? maxRange : turretMax;
             }
             return maxRange;
@@ -246,7 +246,7 @@ namespace PFW.Units.Component.Weapon
         /// Note that this can change as weapons are disabled, and so
         /// return values from this method should not be cached.
         /// </summary>
-        public int MaxRange()
+        public float MaxRange()
         {
             return MaxRange(new TargetTuple(Vector3.zero));
         }
